requirements.txt
beautifulsoup4==4.13.4
MetaTrader5==5.0.5200
pandas==2.3.1
matplotlib==3.10.5
pillow==11.3.0
python-dotenv>=1.0.0
requests==2.32.4
black==25.1.0
flake8>=7.1.1
flake8-bugbear==24.12.12
flake8-pie==0.16.0
isort>=5.9.3
mypy>=1.15.0